:recycle: split up header files
